# build: enable libc++ and libstdc++ hardening