--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -554,6 +554,12 @@ else()
   if(CMAKE_SYSTEM_NAME STREQUAL "Darwin")
     try_append_linker_flag("-Wl,-fixup_chains" TARGET core_interface)
   endif()
+
+  if(HAVE_LIBCPP)
+    # https://libcxx.llvm.org/Hardening.html
+    target_compile_definitions(core_interface INTERFACE
+      $<IF:$<CONFIG:Debug>,_LIBCPP_HARDENING_MODE=_LIBCPP_HARDENING_MODE_DEBUG,>)
+  endif()
 endif()
 
 if(REDUCE_EXPORTS)