--- a/cmake/introspection.cmake
+++ b/cmake/introspection.cmake
@@ -6,6 +6,8 @@ include(CheckCXXSourceCompiles)
 include(CheckCXXSymbolExists)
 include(CheckIncludeFileCXX)
 
+check_cxx_symbol_exists(_LIBCPP_VERSION "version" HAVE_LIBCPP)
+
 check_cxx_symbol_exists(O_CLOEXEC "fcntl.h" HAVE_O_CLOEXEC)
 check_cxx_symbol_exists(fdatasync "unistd.h" HAVE_FDATASYNC)
 check_cxx_symbol_exists(fork "unistd.h" HAVE_DECL_FORK)