Enable the hardening of libc++ and libstdc++.

Inspired by https://github.com/bitcoin/bitcoin/issues/31272#issuecomment-2518700939