@vasild 2025-01-27
`1ef9f7ca4a...8cf75a4a18`: use generator expressions instead of `if(CMAKE_BUILD_TYPE STREQUAL "Debug")` because they also work for the `Ninja Multi-Config` generator.