@fanquake 2025-02-05
What is the status of this? Are you investigating the performance regressions?