@vasild 2025-03-18
The documentation that I linked is explicit:

https://gcc.gnu.org/onlinedocs/libstdc++/manual/using_macros.html

> _GLIBCXX_ASSERTIONS
    Defined by default when compiling with no optimization, undefined by default when compiling with optimization. When defined, enables extra error checking in the form of precondition assertions, such as bounds checking in strings and null pointer checks when dereferencing smart pointers.

If you suspect that there is a bug in GCC or the documentation is wrong, you can try the above `random_tests/t` compiled on `master` (this PR does not change behavior when GCC is used):

`-DCMAKE_BUILD_TYPE=Debug` (uses `-O0`):
```
/usr/local/lib/gcc15/include/c++/bits/stl_vector.h:1262: constexpr std::vector<_Tp, _Alloc>::reference std::vector<_Tp, _Alloc>::operator[](size_type) [with _Tp = int; _Alloc = std::allocator<int>; reference = int&; size_type = long unsigned int]: Assertion '__n < this->size()' failed.
```

`-DCMAKE_BUILD_TYPE=Release` (uses `-O2`):
```
*** No errors detected
```