@vasild 2025-04-02
Can't enable `_GLIBCXX_DEBUG_PEDANTIC` out in the wild (ie for everybody via CMake). But it is already enabled in a controlled environment (the CI): https://github.com/bitcoin/bitcoin/pull/31424#issuecomment-2724631464