@vasild 2025-04-02
> libc++ hardening is also enabled in the CI.

Yes. That is not ABI breaking, so it _can_ be enabled out in the wild (ie for everybody via CMake) in addition to the CI.

> I find it extremely hard to review this pull request, because for months, on every push it is changing what hardening setting to apply in what severity to what part of the system.

I changed it based on reviewers' feedback, including yours. Thanks for that! Also reduced the scope and dropped parts of it because of possible performance issues. Right now it is `+8 -0` lines.

> My recommendation would be to just state your goal, so that reviewers can easily decide what is the best way to achieve your goal.

It is stated in the PR title and description.