@vasild 2025-04-07
> I guess performance expectations are non-existing in debug builds, but it would still be good to know ...

Agree. Debug-performance of the unit tests seems to be the same as in `master`, or within noise:

master:
```
$ for i in {1..5} ; do ctest --test-dir build-master 2>&1 |grep 'Total Test time' ; done
Total Test time (real) = 407.20 sec
Total Test time (real) = 397.09 sec
Total Test time (real) = 422.87 sec
Total Test time (real) = 410.20 sec
Total Test time (real) = 415.94 sec
```

PR:
```
$ for i in {1..5} ; do ctest --test-dir build-pr 2>&1 |grep 'Total Test time' ; done
Total Test time (real) = 403.39 sec
Total Test time (real) = 403.48 sec
Total Test time (real) = 404.27 sec
Total Test time (real) = 394.76 sec
Total Test time (real) = 407.62 sec
```