@fanquake 2025-04-29
`ENABLE_HARDENING` and the `hardening_interface` have been removed.